feat(findYourWorkplaceDirective): Create findYourWorkplace directive and extend in create-account
